Fix warning (unsigned/signed comparison)

+++ src/ClientAuthSession.cpp
@@ -238,7 +238,7 @@ void ClientAuthSession::onPacketServerList(TS_AC_SERVER_LIST& packet) {
 	selectedServer = 0;
 	this->serverList.clear();
 
-	for(int i=0; i < packetServerList.size(); ++i) {
+	for(size_t i = 0; i < packetServerList.size(); ++i) {
 		currentServerInfo.serverId = packetServerList[i].server_idx;
 		currentServerInfo.serverIp = packetServerList[i].server_ip;
 		currentServerInfo.serverPort = packetServerList[i].server_port;
